Replace `#[expect]`/`#[allow]` on selected casts with real handling
Where it's not complicated, prefer a workaround that produces a defined
result for the out-of-range case rather than silently casting:

- `Timestamp::Display`: `i64::try_from(...).unwrap_or(i64::MAX)` for the
  seconds field, and `u32::try_from` (which is always satisfied because
  of the preceding `% 1_000_000`) — no `#[expect]` needed.
- `chain_worker::handle::current_micros`: saturate to `u64::MAX` if the
  duration since UNIX_EPOCH ever overflows.
- `NodeError::InvalidTimestamp.block_time_grace_period_ms`: same
  saturating conversion.
- `RequestsCache` test: drop `cache_size: u64`; pass the literal
  directly and use `0u64..3` for the matching loop.
- `certificate_upload_batch_size`: change the type from `u64` to `usize`
  (it's only used as a chunk size). The sibling
  `sender_certificate_download_batch_size` was already `usize`.
- `linera-storage-service::server`: use `i32::try_from(...).expect(...)`
  for `num_chunks` (silent wrap would corrupt the gRPC reply) and
  `usize::try_from(index).map_err(invalid_argument)` to reject negative
  chunk indices from the wire. Drop the module-level `#![allow]`.
- `*CollectionView::hash`/`hash_mut`: return `ArithmeticError::Overflow`
  via `ViewError` instead of silently truncating the key count fed into
  the hash. Same for the GraphQL `count_` resolvers across collection /
  reentrant_collection / map / set views — they now return a GraphQL
  error if the count exceeds `u32::MAX` rather than wrapping.

+++ linera-core/src/client/chain_client/mod.rs
@@ -109,7 +109,7 @@ pub struct Options {
     pub certificate_download_batch_size: u64,
     /// Maximum number of certificates read from local storage and uploaded to a validator
     /// at a time when synchronizing a chain.
-    pub certificate_upload_batch_size: u64,
+    pub certificate_upload_batch_size: usize,
     /// Maximum number of sender certificates we try to download and receive in one go
     /// when syncing sender chains.
     pub sender_certificate_download_batch_size: usize,
